Changing Job field to 'int64'

+++ src/map/itemdb.h
@@ -468,7 +468,7 @@ struct item_data {
 	int delay;
 //Lupus: I rearranged order of these fields due to compatibility with ITEMINFO script command
 //       some script commands should be revised as well...
-	unsigned int class_base[3]; ///< Specifies if the base can wear this item (split in 3 indexes per type: 1-1, 2-1, 2-2)
+	uint64 class_base[3]; ///< Specifies if the base can wear this item (split in 3 indexes per type: 1-1, 2-1, 2-2)
 	unsigned class_upper : 6;   ///< Specifies if the upper-type can equip it (bitfield, 0x01: normal, 0x02: upper, 0x04: baby normal, 0x08: third normal, 0x10: third upper, 0x20: third baby)
 	struct {
 		unsigned short chance;
@@ -597,8 +597,8 @@ struct itemdb_interface {
 	int (*searchname_array_sub) (DBKey key, DBData data, va_list ap);
 	int (*searchrandomid) (struct item_group *group);
 	const char* (*typename) (int type);
-	void (*jobmask2mapid) (unsigned int *bclass, unsigned int jobmask);
-	void (*jobid2mapid) (unsigned int *bclass, int job_id, bool enable);
+	void (*jobmask2mapid) (uint64 *bclass, int64 jobmask);
+	void (*jobid2mapid) (uint64 *bclass, int job_id, bool enable);
 	void (*create_dummy_data) (void);
 	struct item_data* (*create_item_data) (int nameid);
 	int (*isequip) (int nameid);
